refactor: move job form to more common folder
