+++ apps/web/src/modules/job/components/Jobform/components/ShootingForm/index.tsx
